requirements.txt
PyQt5
pytest-qt